Resto markers get added, but there\'s still a threading problem

+++ android/src/be/ugent/zeus/hydra/data/Resto.java
@@ -10,7 +10,7 @@ public class Resto implements Serializable {
 
     public String name;
     public String address;
-    public Integer longitude;
-    public Integer latitude;
+    public Double longitude;
+    public Double latitude;
     public boolean resto;
 }
